refactor profiler into smaller chunks
